pyproject.toml
[project]
name = "big-data-interactive-ads"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "happybase>=1.3.0",
    "kafka-python>=2.0.2",
    "pyspark>=4.1.0",
    "requests>=2.32.5",
    "pandas>=2.3.3",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "jupyter>=1.1.1",
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "W", "C", "B", "I"]
ignore = ["E203"]
extend-ignore = []
extend-select = []
exclude = ["**/metastore_db/**", "**/.venv/**", "**/venv/**"]
